fix: fix compatibility for Python 2.7, circle import of parser and validator
